// Bottleneko.Client/src/lib/api/dtos.gen.ts
export type ProxyType
    = 'Http'
        | 'Https'
        | 'Socks4'
        | 'Socks4a'
        | 'Socks5';
export const ProxyTypeValues : ProxyType[] = [
    'Http',
    'Https',
    'Socks4',
    'Socks4a',
    'Socks5',
];

export interface ProxyDto {
    id: string;
    name: string;
    type: ProxyType;
    hostname: string;
    port: number;
    isAuthRequired: boolean;
    username: string | null;
    password: string | null;
}